# src/InactivePDF.Api/wwwroot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InactivePDF API",
    "version": "0.1.0-beta",
    "description": "Self-hosted document-to-PDF conversion API. Use an administrator bearer token for administration routes or a scoped integration key for conversion and owned-job routes."
  },
  "servers": [{ "url": "/", "description": "Current InactivePDF instance" }],
...
  "paths": {
...
    "/v1/conversions": { "post": { "tags": ["Conversion"], "summary": "Convert one or more files", "description": "Send multipart fields mode=sync or async, profile, watermarkProfile, and optional JSON watermark. Async returns 202 with a job location; sync streams the PDF.", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "requestBody": { "$ref": "#/components/requestBodies/ConversionMultipart" }, "responses": { "200": { "$ref": "#/components/responses/Pdf" }, "202": { "$ref": "#/components/responses/AcceptedJob" }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" }, "413": { "$ref": "#/components/responses/Problem" }, "422": { "$ref": "#/components/responses/Problem" } } } },
...
    "/v1/jobs/{jobId}": { "get": { "tags": ["Jobs"], "summary": "Read job status", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/JobId" }], "responses": { "200": { "description": "Job status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } }, "404": { "$ref": "#/components/responses/Problem" } } }, "delete": { "tags": ["Jobs"], "summary": "Delete a terminal owned job", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/JobId" }], "responses": { "204": { "description": "Deleted" }, "default": { "$ref": "#/components/responses/Problem" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "administratorBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "administrator token" },
      "integrationBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "integration API key" }
    },
    "parameters": {
      "JobId": { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    },
    "requestBodies": {
      "ConversionMultipart": { "required": true, "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/ConversionForm" } } } },
...
    },
    "schemas": {
...
      "ConversionForm": { "type": "object", "required": ["file"], "properties": { "mode": { "type": "string", "enum": ["sync", "async"], "default": "async" }, "profile": { "type": "string", "default": "archive" }, "watermarkProfile": { "type": "string" }, "watermark": { "type": "string", "description": "JSON-encoded WatermarkOptions for multipart requests. Direct options take precedence over watermarkProfile." }, "file": { "type": "array", "items": { "type": "string", "format": "binary" } } } },
...
      "JobStatus": { "type": "object", "properties": { "jobId": { "type": "string", "format": "uuid" }, "state": { "type": "string" }, "progressPercent": { "type": "integer" }, "queueWaitMilliseconds": { "type": "integer" }, "processingMilliseconds": { "type": "integer" }, "retryCount": { "type": "integer" }, "inputBytes": { "type": "integer", "format": "int64" }, "outputBytes": { "type": "integer", "format": "int64" }, "compressionPercent": { "type": "number", "format": "double" } } },
      "Problem": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" } } }
    },
    "responses": {
      "Pdf": { "description": "Generated PDF", "content": { "application/pdf": { "schema": { "type": "string", "format": "binary" } } } },
      "AcceptedJob": { "description": "Job accepted", "headers": { "Location": { "schema": { "type": "string" } }, "X-Request-ID": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } },
      "Problem": { "description": "Structured error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    }
  }
}